Move AAE code into the aae folder.
